+++ src/components/widgets/index.js
@@ -1,0 +1,5 @@
+import Clock from './Clock.vue';
+import Stardate from './Stardate.vue';
+import Calculator from './Calculator.vue';
+
+export { Clock, Stardate, Calculator };