+++ NEWS.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2020-12-09
+
+- Move documentation to a `vuepress` docs folder with a custom theme
+
 ## 2019-06-24
 
 - Stop using `.lcars-chrome-component` for borders